Архитектура: MVVM от 01.06.2025
