+++ app/src/main/java/ru/netology/nmedia/viewmodel/PostViewModel.kt
@@ -1,0 +1,18 @@
+package ru.netology.nmedia.viewmodel
+
+import androidx.lifecycle.ViewModel
+import ru.netology.nmedia.repository.PostRepository
+import ru.netology.nmedia.repository.PostRepositoryInMemoryImpl
+
+
+// пакет ViewModel отвечает за хранение данных, относящихся к UI, и за
+// связывание UI с бизнес-логикой
+
+
+class PostViewModel: ViewModel(){
+    private val repository: PostRepository = PostRepositoryInMemoryImpl()
+    val data = repository.get()
+    fun like() = repository.like()
+    fun share() = repository.share()
+    fun view() = repository.view()
+}
